# prototype-api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/actions": {
      "post": {
        "operationId": "createAction",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a Action.",
        "tags": [
          "actions"
        ]
      },
      "get": {
        "operationId": "listActions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActionResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List Action records.",
        "tags": [
          "actions"
        ]
      }
    },
    "/actions/{id}": {
      "get": {
        "operationId": "getAction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Action ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Action not found."
          }
        },
        "summary": "Get a Action by ID.",
        "tags": [
          "actions"
        ]
      },
      "patch": {
        "operationId": "updateAction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Action ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Action not found."
          }
        },
        "summary": "Update a Action.",
        "tags": [
          "actions"
        ]
      },
      "delete": {
        "operationId": "deleteAction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Action ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Action deleted."
          },
          "404": {
            "description": "Action not found."
          }
        },
        "summary": "Delete a Action.",
        "tags": [
          "actions"
        ]
      }
    },
    "/changes": {
      "post": {
        "operationId": "createChange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChangeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a Change.",
        "tags": [
          "changes"
        ]
      },
      "get": {
        "operationId": "listChanges",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChangeResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List Change records.",
        "tags": [
          "changes"
        ]
      }
    },
    "/changes/{id}": {
      "get": {
        "operationId": "getChange",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Change ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Change not found."
          }
        },
        "summary": "Get a Change by ID.",
        "tags": [
          "changes"
        ]
      },
      "patch": {
        "operationId": "updateChange",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Change ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChangeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Change not found."
          }
        },
        "summary": "Update a Change.",
        "tags": [
          "changes"
        ]
      },
      "delete": {
        "operationId": "deleteChange",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Change ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Change deleted."
          },
          "404": {
            "description": "Change not found."
          }
        },
        "summary": "Delete a Change.",
        "tags": [
          "changes"
        ]
      }
    },
    "/change-items": {
      "post": {
        "operationId": "createChangeItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChangeItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeItemResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a ChangeItem.",
        "tags": [
          "change-items"
        ]
      },
      "get": {
        "operationId": "listChangeItems",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChangeItemResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List ChangeItem records.",
        "tags": [
          "change-items"
        ]
      }
    },
    "/change-items/{id}": {
      "get": {
        "operationId": "getChangeItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ChangeItem ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeItemResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "ChangeItem not found."
          }
        },
        "summary": "Get a ChangeItem by ID.",
        "tags": [
          "change-items"
        ]
      },
      "patch": {
        "operationId": "updateChangeItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ChangeItem ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChangeItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeItemResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "ChangeItem not found."
          }
        },
        "summary": "Update a ChangeItem.",
        "tags": [
          "change-items"
        ]
      },
      "delete": {
        "operationId": "deleteChangeItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ChangeItem ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "ChangeItem deleted."
          },
          "404": {
            "description": "ChangeItem not found."
          }
        },
        "summary": "Delete a ChangeItem.",
        "tags": [
          "change-items"
        ]
      }
    },
...
  },
  "info": {
    "title": "IFS Prototype API",
    "description": "OpenAPI contract for the IFS prototype backend.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateActionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this action."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Action."
          },
          "entityType": {
            "type": "string",
            "description": "Entity category for this object, normally Action."
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this action record was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this action record was last updated.",
            "format": "date-time"
          },
          "name": {
            "type": "string",
            "description": "Human-readable name of the action."
          },
          "uniqueName": {
            "type": "string",
            "description": "Stable unique name used by IFS implementations to identify this action."
          },
          "requiresPermission": {
            "type": "boolean",
            "description": "Whether a member or role must hold an applicable permission before performing this action."
          },
          "description": {
            "type": "string",
            "description": "Describes the action."
          },
          "instructions": {
            "type": "string",
            "description": "Instructions for how to perform the action."
          },
          "notes": {
            "type": "string",
            "description": "Additional notes about this action."
          },
          "warnings": {
            "type": "string",
            "description": "Warnings, risks, or constraints to review before performing this action."
          },
          "state": {
            "enum": [
              "drafted",
              "under-review",
              "active",
              "suspended",
              "revoked",
              "deprecated"
            ],
            "type": "string",
            "description": "Lifecycle state of the action."
          }
        },
        "required": [
          "id",
          "ifsId",
          "entityType",
          "entityDocumentationUrl",
          "createdAt",
          "updatedAt",
          "name",
          "uniqueName",
          "requiresPermission",
          "description",
          "state"
        ]
      },
      "ActionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this action."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Action."
          },
          "entityType": {
            "type": "string",
            "description": "Entity category for this object, normally Action."
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this action record was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this action record was last updated.",
            "format": "date-time"
          },
          "name": {
            "type": "string",
            "description": "Human-readable name of the action."
          },
          "uniqueName": {
            "type": "string",
            "description": "Stable unique name used by IFS implementations to identify this action."
          },
          "requiresPermission": {
            "type": "boolean",
            "description": "Whether a member or role must hold an applicable permission before performing this action."
          },
          "description": {
            "type": "string",
            "description": "Describes the action."
          },
          "instructions": {
            "type": "string",
            "description": "Instructions for how to perform the action."
          },
          "notes": {
            "type": "string",
            "description": "Additional notes about this action."
          },
          "warnings": {
            "type": "string",
            "description": "Warnings, risks, or constraints to review before performing this action."
          },
          "state": {
            "enum": [
              "drafted",
              "under-review",
              "active",
              "suspended",
              "revoked",
              "deprecated"
            ],
            "type": "string",
            "description": "Lifecycle state of the action."
          }
        },
        "required": [
          "id",
          "ifsId",
          "entityType",
          "entityDocumentationUrl",
          "createdAt",
          "updatedAt",
          "name",
          "uniqueName",
          "requiresPermission",
          "description",
          "state"
        ]
      },
      "UpdateActionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this action."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Action."
          },
          "entityType": {
            "type": "string",
            "description": "Entity category for this object, normally Action."
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this action record was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this action record was last updated.",
            "format": "date-time"
          },
          "name": {
            "type": "string",
            "description": "Human-readable name of the action."
          },
          "uniqueName": {
            "type": "string",
            "description": "Stable unique name used by IFS implementations to identify this action."
          },
          "requiresPermission": {
            "type": "boolean",
            "description": "Whether a member or role must hold an applicable permission before performing this action."
          },
          "description": {
            "type": "string",
            "description": "Describes the action."
          },
          "instructions": {
            "type": "string",
            "description": "Instructions for how to perform the action."
          },
          "notes": {
            "type": "string",
            "description": "Additional notes about this action."
          },
          "warnings": {
            "type": "string",
            "description": "Warnings, risks, or constraints to review before performing this action."
          },
          "state": {
            "enum": [
              "drafted",
              "under-review",
              "active",
              "suspended",
              "revoked",
              "deprecated"
            ],
            "type": "string",
            "description": "Lifecycle state of the action."
          }
        }
      },
...
      "CreateChangeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this change."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Change."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"Change\" for Change entities.",
            "enum": [
              "Change"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "title": {
            "type": "string",
            "description": "Human-readable title summarizing the proposed change."
          },
          "description": {
            "type": "string",
            "description": "Detailed explanation of what this change proposes and why it is needed."
          },
          "summary": {
            "type": "string",
            "description": "Short explanation of what this change proposes and why it is needed."
          },
          "reason": {
            "type": "string",
            "description": "explanation of why this change is being proposed."
          },
          "changes": {
            "description": "ChangeItem entities describing each proposed object-level operation in this change.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChangeItemResponseDto"
            }
          },
          "decision": {
            "description": "Current decision about this change",
            "allOf": [
              {
                "$ref": "#/components/schemas/DecisionResponseDto"
              }
            ]
          },
          "comments": {
            "description": "Recent Comment entities for this change, such as the first page of comments.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentResponseDto"
            }
          },
          "commentsUrl": {
            "type": "string",
            "description": "URL where the full comment thread for this change can be fetched or viewed.",
            "format": "uri"
          },
          "reviewComments": {
            "description": "ReviewComment entities tied to changed objects, fields, records, or diff entries.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewCommentResponseDto"
            }
          },
          "authors": {
            "description": "Member entities that authored or co-authored this change.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberResponseDto"
            }
          },
          "mainAuthorId": {
            "type": "string",
            "description": "Identifier of the primary member responsible for this change."
          },
          "mainAuthor": {
            "description": "Primary member responsible for this change.",
            "allOf": [
              {
                "$ref": "#/components/schemas/MemberResponseDto"
              }
            ]
          },
          "reviewers": {
            "description": "Member entities requested or assigned to review this change.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberResponseDto"
            }
          },
          "status": {
            "enum": [
              "drafted",
              "closed",
              "open",
              "rejected",
              "approved",
              "merged"
            ],
            "type": "string",
            "description": "Lifecycle status of the change."
          },
          "diffUrl": {
            "type": "string",
            "description": "URL showing a diff for all changed objects, similar to a GitHub pull request diff.",
            "format": "uri"
          },
          "number": {
            "type": "number",
            "description": "Human-friendly sequence number for this change inside the relevant system or scope.",
            "minimum": 1
          },
          "labels": {
            "description": "Label assigned to this change.",
            "examples": [
              "main-system:fix",
              "ali:needs-urgent-attention"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "merged": {
            "type": "boolean",
            "description": "Whether this change has been merged into the target system state.",
            "default": false
          },
          "hasConflict": {
            "type": "boolean",
            "description": "Whether this change currently conflicts with another change or with the target system state.",
            "default": false
          },
          "links": {
            "description": "Any links related to this change.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this change was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this change was last updated.",
            "format": "date-time"
          },
          "mergedAt": {
            "type": "string",
            "description": "Optional timestamp when this change was merged.",
            "format": "date-time"
          },
          "closedAt": {
            "type": "string",
            "description": "Optional timestamp when this change was closed without being merged.",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ifsId",
          "entityType",
          "entityDocumentationUrl",
          "title",
          "description",
          "reason",
          "changes",
          "authors",
          "mainAuthorId",
          "mainAuthor",
          "status",
          "number",
          "merged",
          "hasConflict",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}