# prototype-api/src/action/action.controller.ts
// Generated by scripts/generate-resource.mjs.
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  HttpStatus,
  NotFoundException,
  Param,
  Patch,
  Post,
} from '@nestjs/common';
import {
  ApiBody,
  ApiCreatedResponse,
  ApiNoContentResponse,
  ApiNotFoundResponse,
  ApiOkResponse,
  ApiOperation,
  ApiParam,
  ApiTags,
} from '@nestjs/swagger';
import { CreateActionDto } from './dto/create-action.dto.js';
import { ActionResponseDto } from './dto/action-response.dto.js';
import { UpdateActionDto } from './dto/update-action.dto.js';
import { ActionService } from './action.service.js';

@ApiTags('actions')
@Controller('actions')
export class ActionController {
  constructor(private readonly actionService: ActionService) {}

  @Post()
  @ApiOperation({ operationId: 'createAction', summary: 'Create a Action.' })
  @ApiBody({ type: CreateActionDto })
  @ApiCreatedResponse({ type: ActionResponseDto })
  create(@Body() data: CreateActionDto): Promise<ActionResponseDto> {
    return this.actionService.create(data);
  }

  @Get()
  @ApiOperation({ operationId: 'listActions', summary: 'List Action records.' })
  @ApiOkResponse({ type: ActionResponseDto, isArray: true })
  findAll(): Promise<ActionResponseDto[]> {
    return this.actionService.findAll();
  }

  @Get(':id')
  @ApiOperation({ operationId: 'getAction', summary: 'Get a Action by ID.' })
  @ApiParam({ name: 'id', type: String, description: 'Action ID.' })
  @ApiOkResponse({ type: ActionResponseDto })
  @ApiNotFoundResponse({ description: 'Action not found.' })
  async findOne(@Param('id') id: string): Promise<ActionResponseDto> {
    const result = await this.actionService.findOne(id);
    if (!result) throw new NotFoundException('Action not found.');
    return result;
  }

  @Patch(':id')
  @ApiOperation({ operationId: 'updateAction', summary: 'Update a Action.' })
  @ApiParam({ name: 'id', type: String, description: 'Action ID.' })
  @ApiBody({ type: UpdateActionDto })
  @ApiOkResponse({ type: ActionResponseDto })
  @ApiNotFoundResponse({ description: 'Action not found.' })
  update(
    @Param('id') id: string,
    @Body() data: UpdateActionDto,
  ): Promise<ActionResponseDto> {
    return this.actionService.update(id, data);
  }

  @Delete(':id')
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({ operationId: 'deleteAction', summary: 'Delete a Action.' })
  @ApiParam({ name: 'id', type: String, description: 'Action ID.' })
  @ApiNoContentResponse({ description: 'Action deleted.' })
  @ApiNotFoundResponse({ description: 'Action not found.' })
  async remove(@Param('id') id: string): Promise<void> {
    await this.actionService.remove(id);
  }
}