# prototype-api/src/change-item/change-item.controller.ts
// Generated by scripts/generate-resource.mjs.
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  HttpStatus,
  NotFoundException,
  Param,
  Patch,
  Post,
} from '@nestjs/common';
import {
  ApiBody,
  ApiCreatedResponse,
  ApiNoContentResponse,
  ApiNotFoundResponse,
  ApiOkResponse,
  ApiOperation,
  ApiParam,
  ApiTags,
} from '@nestjs/swagger';
import { CreateChangeItemDto } from './dto/create-change-item.dto.js';
import { ChangeItemResponseDto } from './dto/change-item-response.dto.js';
import { UpdateChangeItemDto } from './dto/update-change-item.dto.js';
import { ChangeItemService } from './change-item.service.js';

@ApiTags('change-items')
@Controller('change-items')
export class ChangeItemController {
  constructor(private readonly changeItemService: ChangeItemService) {}

  @Post()
  @ApiOperation({
    operationId: 'createChangeItem',
    summary: 'Create a ChangeItem.',
  })
  @ApiBody({ type: CreateChangeItemDto })
  @ApiCreatedResponse({ type: ChangeItemResponseDto })
  create(@Body() data: CreateChangeItemDto): Promise<ChangeItemResponseDto> {
    return this.changeItemService.create(data);
  }

  @Get()
  @ApiOperation({
    operationId: 'listChangeItems',
    summary: 'List ChangeItem records.',
  })
  @ApiOkResponse({ type: ChangeItemResponseDto, isArray: true })
  findAll(): Promise<ChangeItemResponseDto[]> {
    return this.changeItemService.findAll();
  }

  @Get(':id')
  @ApiOperation({
    operationId: 'getChangeItem',
    summary: 'Get a ChangeItem by ID.',
  })
  @ApiParam({ name: 'id', type: String, description: 'ChangeItem ID.' })
  @ApiOkResponse({ type: ChangeItemResponseDto })
  @ApiNotFoundResponse({ description: 'ChangeItem not found.' })
  async findOne(@Param('id') id: string): Promise<ChangeItemResponseDto> {
    const result = await this.changeItemService.findOne(id);
    if (!result) throw new NotFoundException('ChangeItem not found.');
    return result;
  }

  @Patch(':id')
  @ApiOperation({
    operationId: 'updateChangeItem',
    summary: 'Update a ChangeItem.',
  })
  @ApiParam({ name: 'id', type: String, description: 'ChangeItem ID.' })
  @ApiBody({ type: UpdateChangeItemDto })
  @ApiOkResponse({ type: ChangeItemResponseDto })
  @ApiNotFoundResponse({ description: 'ChangeItem not found.' })
  update(
    @Param('id') id: string,
    @Body() data: UpdateChangeItemDto,
  ): Promise<ChangeItemResponseDto> {
    return this.changeItemService.update(id, data);
  }

  @Delete(':id')
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({
    operationId: 'deleteChangeItem',
    summary: 'Delete a ChangeItem.',
  })
  @ApiParam({ name: 'id', type: String, description: 'ChangeItem ID.' })
  @ApiNoContentResponse({ description: 'ChangeItem deleted.' })
  @ApiNotFoundResponse({ description: 'ChangeItem not found.' })
  async remove(@Param('id') id: string): Promise<void> {
    await this.changeItemService.remove(id);
  }
}