# prototype-api/src/change/change.controller.ts
// Generated by scripts/generate-resource.mjs.
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  HttpStatus,
  NotFoundException,
  Param,
  Patch,
  Post,
} from '@nestjs/common';
import {
  ApiBody,
  ApiCreatedResponse,
  ApiNoContentResponse,
  ApiNotFoundResponse,
  ApiOkResponse,
  ApiOperation,
  ApiParam,
  ApiTags,
} from '@nestjs/swagger';
import { CreateChangeDto } from './dto/create-change.dto.js';
import { ChangeResponseDto } from './dto/change-response.dto.js';
import { UpdateChangeDto } from './dto/update-change.dto.js';
import { ChangeService } from './change.service.js';

@ApiTags('changes')
@Controller('changes')
export class ChangeController {
  constructor(private readonly changeService: ChangeService) {}

  @Post()
  @ApiOperation({ operationId: 'createChange', summary: 'Create a Change.' })
  @ApiBody({ type: CreateChangeDto })
  @ApiCreatedResponse({ type: ChangeResponseDto })
  create(@Body() data: CreateChangeDto): Promise<ChangeResponseDto> {
    return this.changeService.create(data);
  }

  @Get()
  @ApiOperation({ operationId: 'listChanges', summary: 'List Change records.' })
  @ApiOkResponse({ type: ChangeResponseDto, isArray: true })
  findAll(): Promise<ChangeResponseDto[]> {
    return this.changeService.findAll();
  }

  @Get(':id')
  @ApiOperation({ operationId: 'getChange', summary: 'Get a Change by ID.' })
  @ApiParam({ name: 'id', type: String, description: 'Change ID.' })
  @ApiOkResponse({ type: ChangeResponseDto })
  @ApiNotFoundResponse({ description: 'Change not found.' })
  async findOne(@Param('id') id: string): Promise<ChangeResponseDto> {
    const result = await this.changeService.findOne(id);
    if (!result) throw new NotFoundException('Change not found.');
    return result;
  }

  @Patch(':id')
  @ApiOperation({ operationId: 'updateChange', summary: 'Update a Change.' })
  @ApiParam({ name: 'id', type: String, description: 'Change ID.' })
  @ApiBody({ type: UpdateChangeDto })
  @ApiOkResponse({ type: ChangeResponseDto })
  @ApiNotFoundResponse({ description: 'Change not found.' })
  update(
    @Param('id') id: string,
    @Body() data: UpdateChangeDto,
  ): Promise<ChangeResponseDto> {
    return this.changeService.update(id, data);
  }

  @Delete(':id')
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({ operationId: 'deleteChange', summary: 'Delete a Change.' })
  @ApiParam({ name: 'id', type: String, description: 'Change ID.' })
  @ApiNoContentResponse({ description: 'Change deleted.' })
  @ApiNotFoundResponse({ description: 'Change not found.' })
  async remove(@Param('id') id: string): Promise<void> {
    await this.changeService.remove(id);
  }
}